+++ lib/mc-markdown/renderers/html.rb
@@ -1,0 +1,4 @@
+module MCMarkdown
+  class Html < Redcarpet::Render::HTML
+  end
+end