+++ changelog.md
@@ -1,8 +1,9 @@
+- add an html renderer that aliases redcarpet's
+
 # 0.4.0
 - add callout block formatter
 - add a Renderers module, now have a class method to pull memoized instances of renderers
 - reorganize a bit, moved "modules" and renamed them to "formatters"
-
 
 # 0.3.1
 - fix the header_with_id
@@ -15,7 +16,7 @@
 - loosen dependencies
 
 # 0.2.2
-- strip <p> tags surrounding figures
+- strip `<p>` tags surrounding figures
 
 # 0.2.1
 - add ability to specify classes for images/figures